pyproject.toml
[project]
name = "fixmemos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "coloredlogs>=15.0.1",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "lxml>=5.3.2",
    "mistune>=3.1.3",
    "persica>=0.1.0a5",
    "pydantic>=2.11.3",
    "pydantic-settings>=2.8.1",
    "python-dotenv>=1.1.0",
    "starlette>=0.46.2",
    "uvicorn>=0.34.1",
]
